Add files via upload
I just created a new user interface for the employee evaluation screen.  This screen collects values from the user such as (eval_id,emp_id, value1,value2,value3,value4,value5, and score)  The value variable stores the result of the question choice response.  Each result has been coded with a range from 0 to 100 depending on a selection form 1 to 5.  Each question has been weighted individually to be use in the evaluation calculation to generate a raw score.    The results submitted on this screen are pass to the evaluationphp file which is the frontend processing of the results of the evaluation screen.  The primary function is to calculate the raw score, Display on the processing screen, and pass the results to the database.